Configure shutdown mode instead of working around issue.
In the latest JPype release, the shutdown behavior of the Java VM
can be customized. When we configure JPype to skip "destroying" the
JVM, that solves the issue of extended shutdown delays, as well as
issues that emerged from working around that. This was tested to the
extent of running the test suite on Windows and Linux.